+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
     * Loans & Payments
         * [CP-3722] - Hide previously taken input from loan product client eligibility
+        * [CP-3771] - Enable loan end date to be overridden by the loan product configuration
 
     * Tranlations
         * [CP-3741] - Fix Translations Related to Angular Material Pagination and Breadcrumbs